Boton de Geolocalizacion

+++ app/src/main/java/com/example/appweather/router/Ruta.kt
@@ -2,5 +2,5 @@ package com.example.appweather.router
 
 sealed class Ruta(val id: String) {
     data object Ciudades: Ruta("ciudades")
-    data class Clima(val lat: Float,val lon:Float, val nombre:String): Ruta("clima")
+    data class Clima(val lat: Double, val lon: Double, val nombre:String): Ruta("clima")
 }
